refactor(cli): migrate openeo tooling onto discovery
Move openeo's refresher/writer(two-block)/curated_ids/prober/live_validator to
earthlens.openeo.cli; add replace_index_block + bands_from_summaries to the public
toolkit (gee keeps the core bands helper until it migrates). Relocate openeo's CLI
tests to earthlens-imagery. Shrink the guard allow-list (openeo).
